refactor: rename context bar to section filter
